+++ temba/orgs/views.py
@@ -58,17 +58,15 @@ from temba.channels.models import Channel
 from temba.classifiers.models import Classifier
 from temba.flows.models import Flow
 from temba.formax import FormaxMixin
-from temba.utils import analytics, brands, get_anonymous_user, json, languages
+from temba.utils import analytics, get_anonymous_user, json, languages
 from temba.utils.email import is_valid_address, send_template_email
 from temba.utils.fields import (
     ArbitraryJsonChoiceField,
     CheckboxWidget,
-    DateWidget,
     InputWidget,
     SelectMultipleWidget,
     SelectWidget,
     TembaChoiceField,
-    TembaDateTimeField,
 )
 from temba.utils.http import http_headers
 from temba.utils.timezones import TimeZoneFormField
@@ -650,6 +648,7 @@ class UserCRUDL(SmartCRUDL):
         "update",
         "edit",
         "delete",
+        "read",
         "forget",
         "two_factor_enable",
         "two_factor_disable",
@@ -657,10 +656,26 @@ class UserCRUDL(SmartCRUDL):
         "account",
     )
 
+    class Read(StaffOnlyMixin, ContentMenuMixin, SpaMixin, SmartReadView):
+        fields = ("email", "date_joined")
+
+        def build_content_menu(self, menu):
+            menu.add_modax(
+                _("Edit"),
+                "user-update",
+                reverse("orgs.user_update", args=[self.object.id]),
+                title=_("Edit User"),
+                as_button=True,
+            )
+
+            menu.add_modax(
+                _("Delete"), "user-delete", reverse("orgs.user_delete", args=[self.object.id]), title=_("Delete User")
+            )
+
     class List(StaffOnlyMixin, SpaMixin, SmartListView):
-        fields = ("email", "name", "orgs", "date_joined")
+        fields = ("email", "name", "date_joined")
         ordering = ("-date_joined",)
-        search_fields = ("email", "first_name", "last_name")
+        search_fields = ("email__icontains", "first_name__icontains", "last_name__icontains")
         filters = (("all", _("All")), ("beta", _("Beta")), ("staff", _("Staff")))
 
         @csrf_exempt
@@ -670,27 +685,13 @@ class UserCRUDL(SmartCRUDL):
         def get_email(self, user):
             return user.email
 
-        def get_orgs(self, user):
-            orgs = user.get_orgs()[0:6]
-
-            more = ""
-            if len(orgs) > 5:
-                more = ", ..."
-                orgs = orgs[0:5]
-            org_links = ", ".join(
-                [f"<a href='{reverse('orgs.org_update', args=[org.id])}'>{escape(org.name)}</a>" for org in orgs]
-            )
-            return mark_safe(f"{org_links}{more}")
-
         def derive_queryset(self, **kwargs):
             qs = super().derive_queryset(**kwargs).filter(is_active=True).exclude(id=get_anonymous_user().id)
-
             obj_filter = self.request.GET.get("filter")
             if obj_filter == "beta":
                 qs = qs.filter(groups__name="Beta")
             elif obj_filter == "staff":
                 qs = qs.filter(is_staff=True)
-
             return qs
 
         def get_context_data(self, **kwargs):
@@ -699,7 +700,7 @@ class UserCRUDL(SmartCRUDL):
             context["filters"] = self.filters
             return context
 
-    class Update(StaffOnlyMixin, SpaMixin, ComponentFormMixin, ContentMenuMixin, SmartUpdateView):
+    class Update(StaffOnlyMixin, SpaMixin, ModalMixin, ComponentFormMixin, ContentMenuMixin, SmartUpdateView):
         class Form(UserUpdateForm):
             groups = forms.ModelMultipleChoiceField(
                 widget=SelectMultipleWidget(
@@ -716,12 +717,7 @@ class UserCRUDL(SmartCRUDL):
 
         form_class = Form
         success_message = "User updated successfully."
-        title = "Update"
-
-        def build_content_menu(self, menu):
-            menu.add_modax(
-                _("Delete"), "user-delete", reverse("orgs.user_delete", args=[self.object.id]), title=_("Delete User")
-            )
+        title = "Update User"
 
         def pre_save(self, obj):
             obj.username = obj.email
@@ -1726,13 +1722,7 @@ class OrgCRUDL(SmartCRUDL):
             return HttpResponseRedirect(self.get_success_url())
 
     class Plan(InferOrgMixin, OrgPermsMixin, SmartReadView):
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-
-            org = self.request.org
-            context["plan"] = org.plan
-            context["plan_end"] = org.format_datetime(org.plan_end, show_time=False) if org.plan_end else None
-            return context
+        pass
 
     class WhatsappCloudConnect(InferOrgMixin, OrgPermsMixin, SmartFormView):
         class WhatsappCloudConnectForm(forms.Form):
@@ -2013,6 +2003,7 @@ class OrgCRUDL(SmartCRUDL):
                 reverse("orgs.org_update", args=[obj.id]),
                 title=_("Edit Workspace"),
                 as_button=True,
+                on_submit="handleWorkspaceUpdated()",
             )
 
             if obj.is_flagged:
@@ -2049,22 +2040,12 @@ class OrgCRUDL(SmartCRUDL):
                     users_roles.append(dict(role_display=role.display_plural, users=role_users))
 
             context["users_roles"] = users_roles
-            workspaces = Org.objects.filter(parent=org, is_active=True).order_by("-created_on", "name")
-
-            children = {}
-            for workspace in workspaces:
-                plan_workspaces = children.get(workspace.plan, [])
-                plan_workspaces.append(workspace)
-                children[workspace.plan] = plan_workspaces
-
-            context["children"] = children
-
+            context["children"] = Org.objects.filter(parent=org, is_active=True).order_by("-created_on", "name")
             return context
 
     class Manage(StaffOnlyMixin, SpaMixin, SmartListView):
-        fields = ("name", "owner", "created_on", "service")
-        field_config = {"service": {"label": ""}}
-        default_order = ("-created_on",)
+        fields = ("name", "owner", "timezone", "created_on")
+        default_order = ("plan_end", "-created_on")
         search_fields = ("name__icontains", "created_by__email__iexact", "config__icontains")
         link_fields = ("name", "owner")
         filters = (
@@ -2081,39 +2062,31 @@ class OrgCRUDL(SmartCRUDL):
 
         def get_owner(self, obj):
             owner = obj.get_owner()
+            return f"{owner.name} ({owner.email})"
 
-            return mark_safe(
-                f"<div class='owner-name'>{escape(owner.first_name)} {escape(owner.last_name)}</div><div class='owner-email'>{escape(owner.username)}</div>"
-            )
+        def derive_queryset(self, **kwargs):  # pragma: no cover
+            obj_filter = self.request.GET.get("filter")
 
-        def get_service(self, obj):
-            url = reverse("orgs.org_service")
-
-            return mark_safe(
-                "<div onclick='goto(event)' href='%s?organization=%d' class='service posterize hover-linked text-gray-400'><temba-icon name='icon.service'></div></div>"
-                % (url, obj.id)
-            )
-
-        def get_name(self, obj):
-            return mark_safe(f"<div class='org-name'>{escape(obj.name)}</div><div class='org-plan'>{obj.plan}</div>")
-
-        def derive_queryset(self, **kwargs):
             qs = super().derive_queryset(**kwargs).filter(is_active=True)
-
             qs = qs.filter(brand=self.request.branding["slug"])
 
-            obj_filter = self.request.GET.get("filter")
             if obj_filter == "anon":
-                qs = qs.filter(is_anon=True)
+                qs = qs.filter(is_anon=True, is_suspended=False)
             elif obj_filter == "flagged":
-                qs = qs.filter(is_flagged=True)
+                qs = qs.filter(is_flagged=True, is_suspended=False)
             elif obj_filter == "suspended":
                 qs = qs.filter(is_suspended=True)
             elif obj_filter == "verified":
                 # this is not my favorite
-                qs = qs.filter(config__icontains='"verified": True')
-            elif obj_filter and obj_filter != "all":
-                qs = qs.filter(Q(plan=obj_filter) | Q(name__icontains=obj_filter))
+                qs = qs.filter(config__icontains='"verified": True', is_suspended=False)
+            elif obj_filter:
+                qs = (
+                    qs.filter(parent=None)
+                    .filter(Q(plan=obj_filter) | Q(name__icontains=obj_filter))
+                    .filter(is_suspended=False)
+                )
+            else:
+                qs = qs.filter(is_suspended=False)
 
             return qs
 
@@ -2129,14 +2102,8 @@ class OrgCRUDL(SmartCRUDL):
                 return reverse("orgs.user_update", args=[owner.pk])
             return super().lookup_field_link(context, field, obj)
 
-        def get_created_by(self, obj):  # pragma: needs cover
-            return "%s %s - %s" % (obj.created_by.first_name, obj.created_by.last_name, obj.created_by.email)
-
     class Update(StaffOnlyMixin, SpaMixin, ModalMixin, ComponentFormMixin, SmartUpdateView):
         class Form(forms.ModelForm):
-            brand = forms.ChoiceField(choices=brands.get_choices(), label=_("Brand"), required=True)
-            parent = forms.IntegerField(required=False)
-            plan_end = TembaDateTimeField(label=_("Plan End Date"), required=False)
             features = forms.MultipleChoiceField(
                 choices=Org.FEATURES_CHOICES, widget=SelectMultipleWidget(), required=False
             )
@@ -2144,15 +2111,8 @@ class OrgCRUDL(SmartCRUDL):
             def __init__(self, org, *args, **kwargs):
                 super().__init__(*args, **kwargs)
 
-                self.fields["plan_end"].widget = DateWidget(attrs={"timezone": org.timezone, "time": True})
-
                 self.limits_rows = []
                 self.add_limits_fields(org)
-
-            def clean_parent(self):
-                parent = self.cleaned_data.get("parent")
-                if parent:
-                    return Org.objects.filter(pk=parent).first()
 
             def clean(self):
                 super().clean()
@@ -2183,10 +2143,6 @@ class OrgCRUDL(SmartCRUDL):
                 model = Org
                 fields = (
                     "name",
-                    "brand",
-                    "parent",
-                    "plan",
-                    "plan_end",
                     "features",
                     "is_anon",
                     "is_suspended",
@@ -2194,6 +2150,7 @@ class OrgCRUDL(SmartCRUDL):
                 )
 
         form_class = Form
+        success_url = "hide"
 
         def derive_title(self):
             return None
@@ -2202,9 +2159,6 @@ class OrgCRUDL(SmartCRUDL):
             kwargs = super().get_form_kwargs()
             kwargs["org"] = self.get_object()
             return kwargs
-
-        def get_success_url(self):
-            return reverse("orgs.org_read", args=[self.get_object().pk])
 
         def post(self, request, *args, **kwargs):
             if "action" in request.POST: