+++ temba/settings_common.py
@@ -259,18 +259,11 @@ PERMISSIONS_APP = "temba.airtime"
 LOGGING = {
     "version": 1,
     "disable_existing_loggers": True,
-    "root": {"level": "WARNING", "handlers": ["console"]},
-    "formatters": {"verbose": {"format": "%(levelname)s %(asctime)s %(module)s %(process)d %(thread)d %(message)s"}},
+    "formatters": {"verbose": {"format": "%(levelname)s %(asctime)s %(module)s %(message)s"}},
     "handlers": {
         "console": {"level": "DEBUG", "class": "logging.StreamHandler", "formatter": "verbose"},
-        "null": {"class": "logging.NullHandler"},
     },
-    "loggers": {
-        "pycountry": {"level": "ERROR", "handlers": ["console"], "propagate": False},
-        "django.security.DisallowedHost": {"handlers": ["null"], "propagate": False},
-        "django.db.backends": {"level": "ERROR", "handlers": ["console"], "propagate": False},
-        "temba.formax": {"level": "DEBUG" if DEBUG else "ERROR", "handlers": ["console"], "propagate": False},
-    },
+    "root": {"level": "INFO", "handlers": ["console"]},
 }
 
 # -----------------------------------------------------------------------------------
@@ -713,7 +706,6 @@ GROUP_PERMISSIONS = {
         "contacts.contactfield_api",
         "contacts.contactfield_read",
         "contacts.contactgroup_api",
-        "contacts.contactgroup_list",
         "contacts.contactgroup_menu",
         "contacts.contactgroup_read",
         "contacts.contactimport_read",
@@ -858,6 +850,7 @@ CACHES = {
 # Async tasks using Celery
 # -----------------------------------------------------------------------------------
 CELERY_RESULT_BACKEND = None
+CELERY_TASK_TRACK_STARTED = True
 CELERY_BROKER_URL = "redis://%s:%d/%d" % (REDIS_HOST, REDIS_PORT, REDIS_DB)
 
 # by default, celery doesn't have any timeout on our redis connections, this fixes that